reworking tests to match migrated and renamed code for surface nets
