+++ Sources/Voxels/Renderers/VoxelMeshRenderer.swift
@@ -1,1 +1,4 @@
-public class VoxelMeshRenderer {}
+public class VoxelMeshRenderer {
+    /// This stride of the SDF array did not produce a vertex.
+    static let NULL_VERTEX = UInt32.max
+}
